Refactor project structure by splitting App class and organizing files
