+++ ElevatorChallenge/src/Models/BuildingConfig.cs
@@ -1,0 +1,8 @@
+﻿namespace ElevatorChallenge.ElevatorChallenge.src.Models
+{
+    public class BuildingConfig
+    {
+        public int TotalFloors { get; set; }
+        // public List<ElevatorConfig> Elevators { get; set; }
+    }
+}
